#45 refactor: ext 파일을 common 폴더로 이동
